basic backend api runable in docker and does response to some requests
